Development

+++ src/main/java/br/com/felipesoarestech/api/cliente/domain/dto/ProductRequestDTO.java
@@ -8,6 +8,6 @@ public record ProductRequestDTO(
         String name,
 
         @NotNull
-        Integer price
+        Double price
 ) {
 }
